+++ source/extensions/clusters/redis/redis_cluster.h
@@ -278,7 +278,8 @@ public:
 private:
   friend class RedisClusterTest;
 
-  Upstream::ClusterImplBaseSharedPtr createClusterWithConfig(
+  std::pair<Upstream::ClusterImplBaseSharedPtr, Upstream::ThreadAwareLoadBalancerPtr>
+  createClusterWithConfig(
       const envoy::api::v2::Cluster& cluster,
       const envoy::config::cluster::redis::RedisClusterConfig& proto_config,
       Upstream::ClusterFactoryContext& context,